sfs-recoll: create custom dired-next-line commands to display file info
Ugh, finally. Couldn't get this working the way I wanted in ivy, so
created a custom minibuffer command. Canceling the previous minibuffer
before the next command to prevent accumulation of recursive minibuffers
was complicated. It required use of timers because exit-minibuffer does a
throw internally so it leaves the context.