sfs-recoll: lookup properties based on md5
This allows files to be looked up in any dired buffer if it is in the
index, and allows the removal of the filename lookup table. The downside
is files can't be looked up directly based on filename only because
Recoll doesn't store the full url string. So if the file changes in any
way without recoll knowing, it won't find the entry. So as a backup, I
at least try to infer if the file is there by looking up an
approximation of the full path, but it could be broken by paths with the
same strings but different orders.